API overhaul
