+++ healvis/tests/test_observatory.py
@@ -1,4 +1,4 @@
-from healvis import visibility, skymodel
+from healvis import observatory, sky_model, beam_model, utils
 from astropy.cosmology import WMAP9
 import nose.tools as nt
 import numpy as np
@@ -15,6 +15,57 @@ latitude = -30.7215277777
 longitude = 21.4283055554
 
 
+def test_Observatory():
+    # TODO: fill out test coverage
+
+    # setup
+    ant1 = np.array([15.0, 0, 0])
+    ant2 = np.array([0.0, 0, 0])
+    bl = observatory.Baseline(ant1, ant2)
+    Npix, Nfreqs = 10, 100
+    az = np.linspace(0, 2 * np.pi, Npix)
+    za = np.linspace(0, np.pi, Npix)
+    freqs = np.linspace(100e6, 200e6, Nfreqs)
+    obs = observatory.Observatory(latitude, longitude, array=[bl], freqs=freqs)
+
+    # test Analytic set beam
+    for beam in ['uniform', 'gaussian', 'airy', beam_model.airy_disk]:
+        obs.set_beam(beam, sigma=10, diameter=10)
+        nt.assert_true(isinstance(obs.beam, beam_model.AnalyticBeam))
+        b = obs.beam.beam_val(az, za, freqs, pol='xx')
+        nt.assert_equal(b.shape, (Npix, Nfreqs))
+
+    # test Power set beam
+    obs.set_beam(os.path.join(DATA_PATH, "HERA_NF_dipole_power.beamfits"))
+    nt.assert_true(isinstance(obs.beam, beam_model.PowerBeam))
+    b = obs.beam.beam_val(az, za, freqs, pol='xx')
+    nt.assert_equal(b.shape, (Npix, Nfreqs))
+
+
+def test_Baseline():
+    # TODO: fill out test coverage
+
+    # setup
+    ant1 = np.array([15.0, 0, 0])
+    ant2 = np.array([0.0, 0, 0])
+    bl = observatory.Baseline(ant1, ant2)
+    Npix, Nfreqs = 10, 100
+    az = np.linspace(0, 2 * np.pi, Npix)
+    za = np.linspace(0, np.pi, Npix)
+    freqs = np.linspace(100e6, 200e6, Nfreqs)
+
+    # test fringe is right dimensions
+    fringe = bl.get_fringe(az, za, freqs)
+    nt.assert_equal(fringe.shape, (Npix, Nfreqs))
+
+    # test fringe at zenith is 1.0 across all freqs
+    nt.assert_true(np.isclose(fringe[0, :], 1.0).all())
+
+
+####################
+# Validation Tests #
+####################
+
 @nt.nottest
 def test_pointings():
     t0 = 2451545.0  # Start at J2000 epoch
@@ -22,7 +73,7 @@ def test_pointings():
     dt_days = dt_min * 1 / 60. * 1 / 24.  # 20 minutes in days
 
     time_arr = np.arange(20) * dt_days + t0
-    obs = visibility.Observatory(latitude, longitude)
+    obs = observatory.Observatory(latitude, longitude)
 
     obs.set_pointings(time_arr)
 
@@ -41,7 +92,7 @@ def test_az_za():
     Check the calculated azimuth and zenith angle of a point exactly 5 deg east on the sphere (az = 90d, za = 5d)
     """
     Nside = 128
-    obs = visibility.Observatory(latitude, longitude)
+    obs = observatory.Observatory(latitude, longitude)
     center = [0, 0]
     lon, lat = [5, 0]
     ind0 = hp.ang2pix(Nside, lon, lat, lonlat=True)
@@ -64,7 +115,7 @@ def test_vis_calc():
     ant1_enu = np.array([0, 0, 0])
     ant2_enu = np.array([0.0, 14.6, 0])
 
-    bl = visibility.Baseline(ant1_enu, ant2_enu)
+    bl = observatory.Baseline(ant1_enu, ant2_enu)
 
     freqs = np.array([1e8])
     nfreqs = 1
@@ -81,15 +132,15 @@ def test_vis_calc():
     shell = np.zeros((npix, nfreqs))
     pix_area = 4 * np.pi / float(npix)
     shell[ind] = 1  # Jy/pix
-    shell[ind] *= visibility.jy2Tstr(freqs[0], bm=pix_area)  # K
+    shell[ind] *= utils.jy2Tstr(freqs[0], bm=pix_area)  # K
 
-    obs = visibility.Observatory(latitude, longitude, array=[bl], freqs=freqs)
+    obs = observatory.Observatory(latitude, longitude, array=[bl], freqs=freqs)
     obs.pointing_centers = centers
     obs.times_jd = np.array([1])
     obs.set_fov(fov)
     obs.set_beam('uniform')
 
-    sky = skymodel.SkyModel(Nside=nside, freq_array=freqs, data=shell)
+    sky = sky_model.SkyModel(Nside=nside, freq_array=freqs, data=shell)
 
     visibs, times, bls = obs.make_visibilities(sky)
     print(visibs)
@@ -106,7 +157,7 @@ def test_offzenith_vis():
     ant1_enu = np.array([0, 0, 0])
     ant2_enu = np.array([0.0, 140.6, 0])
 
-    bl = visibility.Baseline(ant1_enu, ant2_enu)
+    bl = observatory.Baseline(ant1_enu, ant2_enu)
 
     Nside = 128
     ind = 9081
@@ -120,16 +171,16 @@ def test_offzenith_vis():
     phi, theta = hp.pix2ang(Nside, ind, lonlat=True)
     ind = hp.ang2pix(Nside, phi, theta - 5, lonlat=True)
     shell[ind] = 1  # Jy/pix
-    shell[ind] *= visibility.jy2Tstr(freqs[0], pix_area)  # K
+    shell[ind] *= utils.jy2Tstr(freqs[0], pix_area)  # K
 
-    obs = visibility.Observatory(latitude, longitude, array=[bl], freqs=freqs)
+    obs = observatory.Observatory(latitude, longitude, array=[bl], freqs=freqs)
     obs.pointing_centers = [[phi, theta]]
     obs.times_jd = np.array([1])
     obs.set_fov(fov)
     resol = np.sqrt(pix_area)
     obs.set_beam('uniform')
 
-    sky = skymodel.SkyModel(Nside=Nside, freq_array=np.array(freqs), data=shell)
+    sky = sky_model.SkyModel(Nside=Nside, freq_array=np.array(freqs), data=shell)
 
     vis_calc, times, bls = obs.make_visibilities(sky)
 
@@ -146,75 +197,3 @@ def test_offzenith_vis():
     print(vis_calc)
 
     nt.assert_true(np.isclose(vis_calc, vis_analytic, atol=1e-3).all())
-
-
-def test_PowerBeam():
-    # load it
-    beam_path = os.path.join(DATA_PATH, "HERA_NF_dipole_power.beamfits")
-    P = visibility.PowerBeam(beam_path)
-    freqs = np.arange(120e6, 160e6, 4e6)
-    Nfreqs = len(freqs)
-
-    # test frequency interpolation
-    P2 = copy.deepcopy(P)
-    P3 = P2.interp_freq(freqs, inplace=False, kind='linear')
-    P2.interp_freq(freqs, inplace=True, kind='linear')
-    # assert inplace and not inplace are consistent
-    nt.assert_equal(P2, P3)
-    # assert correct frequencies
-    np.testing.assert_array_almost_equal(freqs, P3.freq_array[0])
-    nt.assert_true(P3.bandpass_array.shape[1] == P3.Nfreqs == Nfreqs)
-
-    # get beam value
-    Npix = 20
-    az = np.linspace(0, 2 * np.pi, Npix, endpoint=False)
-    za = np.linspace(0, 1, Npix, endpoint=False)
-    b = P.beam_val(az, az, freqs, pol='XX')
-    # check shape and rough value check (i.e. interpolation is near zenith as expected)
-    nt.assert_equal(b.shape, (Npix, Nfreqs))
-    nt.assert_true(np.isclose(b.max(), 1.0, atol=1e-3))
-
-    # shift frequnecies by a delta and assert beams are EXACLTY the same (i.e. no freq interpolation)
-    # delta must be larger than UVBeam._inter_freq tol, but small enough
-    # to keep the same freq nearest neighbors
-    b2 = P.beam_val(az, az, freqs + 1e6, pol='XX')
-    np.testing.assert_array_almost_equal(b, b2)
-
-
-def test_AnalyticBeam():
-    freqs = np.arange(120e6, 160e6, 4e6)
-    Nfreqs = len(freqs)
-    Npix = 20
-    az = np.linspace(0, 2 * np.pi, Npix, endpoint=False)
-    za = np.linspace(0, 1, Npix, endpoint=False)
-
-    # Gaussian
-    A = visibility.AnalyticBeam('gaussian', sigma=15.0)
-    b = A.beam_val(az, za, freqs)
-    nt.assert_equal(b.shape, (Npix, Nfreqs))  # assert array shape
-    nt.assert_true(np.isclose(b[0, :], 1.0).all())  # assert peak normalized
-
-    # Uniform
-    A = visibility.AnalyticBeam('uniform')
-    b = A.beam_val(az, za, freqs)
-    nt.assert_equal(b.shape, (Npix, Nfreqs))  # assert array shape
-    nt.assert_true(np.isclose(b, 1.0).all())
-
-    # Airy
-    A = visibility.AnalyticBeam('airy', diameter=15.0)
-    b = A.beam_val(az, za, freqs)
-    nt.assert_equal(b.shape, (Npix, Nfreqs))  # assert array shape
-    nt.assert_true(np.isclose(b[0, :], 1.0).all())  # assert peak normalized
-
-    # custom
-    A = visibility.AnalyticBeam(visibility.airy_disk)
-    b2 = A.beam_val(az, za, freqs, diameter=15.0)
-    nt.assert_equal(b2.shape, (Npix, Nfreqs))  # assert array shape
-    nt.assert_true(np.isclose(b2[0, :], 1.0).all())  # assert peak normalized
-    np.testing.assert_array_almost_equal(b, b2)  # assert its the same as airy
-
-    # exceptions
-    A = visibility.AnalyticBeam("uniform")
-    nt.assert_raises(NotImplementedError, visibility.AnalyticBeam, "foo")
-    nt.assert_raises(KeyError, visibility.AnalyticBeam, "gaussian")
-    nt.assert_raises(KeyError, visibility.AnalyticBeam, "airy")
