+++ healvis/utils.py
@@ -3,6 +3,10 @@ from __future__ import absolute_import, division, print_function
 import numpy as np
 import multiprocessing as mp
 from astropy.cosmology import Planck15 as cosmo
+from astropy.constants import c
+
+
+c_ms = c.to('m/s').value
 
 
 class mparray(np.ndarray):
@@ -19,6 +23,18 @@ class mparray(np.ndarray):
         arr = mp.RawArray(ctype, size)
         self.data = arr
         self.reshape(self.shape)
+
+
+def jy2Tstr(f, bm=1.0):
+    '''Return [K sr] / [Jy] vs. frequency (in Hz)
+        Arguments:
+            f = frequencies (Hz)
+            bm = Reference area (defaults to 1 steradian)
+    '''
+    c_cmps = c_ms * 100.   # cm/s
+    k_boltz = 1.380658e-16   # erg/K
+    lam = c_cmps / f  # cm
+    return 1e-23 * lam**2 / (2 * k_boltz * bm)
 
 
 def comoving_voxel_volume(z, dnu, omega):
