+++ scripts/vis_calc.py
@@ -17,7 +17,7 @@ and save to MIRIAD file
 """
 
 import numpy as np
-from healvis import visibility, utils
+from healvis import observatory, utils
 import pylab as pl
 from scipy.stats import binned_statistic
 import os, sys, yaml
@@ -51,7 +51,7 @@ altitude = 1073.
 fov = args.fov  #Deg
 ant1_enu = np.array([0, 0, 0])
 ant2_enu = np.array([0.0, args.bllen , 0])
-bl = visibility.baseline(ant1_enu, ant2_enu)
+bl = observatory.Baseline(ant1_enu, ant2_enu)
 
 # Time
 
@@ -90,7 +90,7 @@ visibs = []
 #fwhms = [35, 40, 45.0, 50.0, 55.0, 60.0]
 fwhm = args.fwhm
 sigma = fwhm/2.355
-obs = visibility.observatory(latitude, longitude, array=[bl], freqs=freqs)
+obs = observatory.Observatory(latitude, longitude, array=[bl], freqs=freqs)
 obs.set_fov(fov)
 obs.set_pointings(time_arr)
 obs.set_beam('gaussian', sigma=sigma)
